-[TNSValueWrapper value] can return null
When memory-managing an instance of a class derived from an Objective-C one there is that sliver of time after the JavaScript representation of the object has been garbage-collected, but the Objective-C representation is still alive. Calls to `retain` or `release` will try to access a JavaScript object that's not there anymore, so guard against that.