Always invoke build script for WebKit and MetadataGenerator targets
They manage their dependencies correctly and do not rebuild anything when no changes are made.
With the current scheme these are being built only once after the project is generated and
never again until the timestamps files are removed. If you change anything or have another reason
to rebuild, you have to manually delete the timestamps.