fix(Inspector): Revert to calling native method in FFICall::async outside of the VM lock
Native call is made outside of the VM lock by design.
For more information see https://github.com/NativeScript/ios-runtime/issues/215 and it's corresponding PR.
This creates a racing condition which might corrupt the internal state of the VM but
a fix for it is outside of this PR's scope, so I'm leaving it as it has always been till now.

We could address the issue by reimplementing the feature with a new VM or a JSWorker.