fix(inspector): Close debugging port when it's not used
Schedule check for existing frontend client 30 sec. after opening socket or disconnecting the frontend.
This is necessary because all {N} apps use one and the same port for debugging communication (18181).
If we don't free it as soon as we're finished with the debugger, no other {N} app will be able to do
LiveSync or be debugged on the same device.