fix(debugging): Don't close inspector socket after 30 sec. timeout
We had to close it in order to have the hardcoded port number 18181
available for other apps after finishing (or never starting) a
debugging session. Without this timeout the first app that opened
the port would take it as long as it's running and would prevent
any subsequent debugging attempts of another app on the same iOS device;
or even the same app on another iOS Simulator on the same Mac machine.

From now on, if the 18181 port is unavailable, a random free port \
will be used and its number will be output in the application log.
When starting a debug session, {N} CLI will monitor the logs
in order to receive the allocated port number for the current
session.