fix(debugging): Change default debugging port to 18183
It turned out that 18182 is used by {N} CLI. It is specified as the port
to listen when launching adb server. This lead to a mysterious bug
when both android and ios devices were connected. To avoid it
for users with previous CLIs we'll change it to 18183.

For more details see:
https://github.com/NativeScript/nativescript-cli/pull/3614