fix(debugging): Do not clear inspector if it doesn't match failed connection
In the scenario when a new connection disconnects the previous one there's
a racing condition which sometimes leads to disconnecting both frontends.
To avoid clearing the new one when the older receives EABORT, we now
check whether the current global inspector_io matches the one of the
failed connection. This way even if the order of events is unfavoring we
won't disconnect the wrong frontend client.

Additional fix for #927