fix(bridge): Don't crash when returned value is not derived from NSObject
There are cases when calling an ObjectiveC method doesn't return a valid NSObject:

* When a JS implementation throws a JS error it has no control to what the return
value of the ObjectiveC method will be. It's always some random memory.
* When an ObjectiveC method returns a C primitive type (e.g. int, long,void, etc.)

Whenever this happens the runtime crashes trying to send messages to them
(e.g. when calling `isKindOfClass` in `JSValue toValue(ExecState* execState, id object, Class klass)`)
In order to protect from such crashes when the returned object is not detected to be a valid
ObjectiveC object, it's value is set to `nil`.

The heuristics algorithm for detecting ObjectiveC objects is taken from
https://blog.timac.org/2016/1124-testing-if-an-arbitrary-pointer-is-a-valid-objective-c-object/.