feat: Copy returned vectors in memory and pass isStructMember to finishCreation
* As cif's invocation is deallocated after the ffi_call, we need to
copy the buffer so we don't pass invalid memory location to js.

* ExtVectorTypeInstance::finishCreation method needs to know whether
or not the vector passed is a struct member because vectors of doubles
are treated differently. For example - when a double3 vector is a
struct member it needs to be aligned as a 4-element vector. When it is
returned independently though - the third element is returned in
ST0 register.