fix(moduleLoader): Don't cache string which eligible for GC
In the promise chain of `commonJSRequire` we need to store the key of the module for
use in а native continuation. The previous approach however stored the `JSValue` which
is wraps a pointer to the underlying `JSString` object. The object however is taken from the
continuation's argument #0 which is eligible for GC and as consequence from time to time
becomes replaced with some other valid string. This behaviour was steadily reproduced
with the release build of NativeScript.framework for ARM64.