fix(webkit): Optimize exception scope in release builds
We've enabled CatchScope EXCEPTION_SCOPE_VERIFICATION in release builds
of NativeScript because we need to know the recursion depth of exceptions. This
however turned out to be a rather hot execution path and leads to significant performance
degradation of the JSC engine (deltaBlue execution is being slowed down somewhere
between 10 and 20% because about 230,000 exception scopes are being created and
destroyed during a single iteration of the algorithm).

To optimize execution we leave only the depth tracking portion of the code when
NDEBUG is defined and move CatchScope, ExceptionScope and ThrowScope classes'
method implementations to header files as ALWAYS_INLINE functions.