fix(metadata-typings): Accept Pointers and References for CString args
Native functions accepting `char *` arguments may need them not only
for receiving a string but also for returning one in a user provided buffer.

In order to not require TS developers to cast `interop.alloc`-ed
buffers in such calls we are starting to generate the typings so that
it's allowed.

See https://github.com/NativeScript/ios-runtime/issues/1017#issuecomment-440968998 for an
example where such a cast was needed.