feat(bridge): Provide automatic garbage collection triggering
Introduce settings similar to the ones in {N} Android Runtime for
configuring automatic GC triggering by the runtime. They are
specified in `app/package.json` as values in the `ios` section:

* `gcThrottleTime` - number of milliseconds which must have passed
since the last automatic GC in order to trigger one during a call from
JS to native
* `memoryCheckInterval` - number of milliseconds which must have
passed since the last automatic GC in order to check the memory
usage on the device and trigger GC if free memory is below the specified
threshold
* `freeMemoryRatio` - a floating-point value from 0 to 1 which sets the
threshold below which a GC will be triggered when the above check is made.

implements #1035