fix(runtime): Keep JS instances in Strong references after creation
Runtime's stability is compromised when automatic GC triggering is
set to a very small interval (e.g. 1 ms). It turns out that if a GC passes
between the moment when an object is created and it's assigned to a
variable, it can get destroyed prematurely. In order to avoid it we have to
keep such objects as `Strong` references.