fix(TestRunner): Turn on automatic GC
Set the automatic triggering period to 5 ms to keep the runtime
and it's tests working correctly with frequent GCs.

If sporadic crashes related to GCs start happening we may set this
value to something significantly lower (e.g. 1 ns = `1e-6` ms)
to instruct the runtime to trigger a new GC on each native call.
This isn't the default value because it significantly degrades
performance.