fix(bridge): Marshal arguments of async calls in worker thread
Temporary ObjectiveC objects which are created during marshalling
are autorelease-d. This means that they are destroyed when the nearest
autorelease pool is drained. Due to the refactoring in this PR, the racing
condition between the JS calling thread for destroying them and the worker
thread using them when performing the method call is won by the JS thread.

To fix this we now marshal the arguments in the worker thread and reject the
promise if something goes wrong instead of doing it synchronously and throwing
an exception.