fix(interop): Add methods `toHexString` and `toDecimalString` to Pointer
This allows for correct serialization of pointers through strings. Which can be
used for passing native objects to workers as discussed in https://github.com/NativeScript/ios-runtime/issues/620#issuecomment-449045591.
The need for these methods arises from the fact that `toNumber` can't
be used for negative pointer values, due to JavaScript's values
inability to represent very large and/or negative 64-bit integer values
as integers.

There is a real-world use-case for the value of `-1` shown in #921