fix: fix waitForDebugger handling
After the 5.1.1 releases of the CLI and the Runtime, the CLI is always sending `AttachRequest` before a debug session which is clearing the `listenSource`. In this way, the app is resumed and the `--debug-brk` flag is not working properly. 

In order to avoid that, we do not resume the app when closing the listening source anymore. It will be resumed only when a debugging client is attached or the 30 secs timeout is reached.