fix(bridge): Remove ObjCWrapperObject from cache after dealloc
When `dealloc` is implemented in JS we were caching a wrapper to the
already deallocated instance. This doesn't do any harm until the same
memory address is reused for another Objective-C instance. When this happens
the new object will be returned to the JS world via the cached wrapper,
which however won't increment its retention count, leading to its premature
deallocation and a crash whenever it's attempted to be used afterwards.