fix(inspector): Catch and log inspector_modules.js loading errors
`inspector_modules.js` is loaded whenever a new debugger connection is established.
When the application is built with `--bundle` however, the file is missing and debugging
is not working correctly. Log the rejection reason to the console so that errors with it don't
silently remain undetected in the future.

related to #1054, #1055, #1056, #1057