fix(webkit): Make JSC log to system logs instead of stderr
Since the implementation of #1064 it has been made possible
for {N} users to switch on different JSC diagnostic options.
The problem was that to view the logs users had to use the Xcode
debugger, because they weren't visible in the system logs and
could not be captured by {N} CLI. With this change now all
applications that retrieve system logs from the device will process
them.

We will also be able to add automated tests that ensure
these options' correct functionality.

related to #1004