fix(build): Fix WebKit project generation errors with CMake 3.14
It turns out that with recent versions of CMake the C++ feature tests
project fails to build correctly when some of the environment variables
inherited by the Xcode build of the NativeScript project are present.
Clear the environment before starting the subproject build.

The variables affecting the build are:
```
SDKROOT="/Applications/Xcode.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS12.1.sdk"
CODE_SIGNING_REQUIRED="YES"
```
If any of them is present the environment, the configuration step of WebKit
fails with error `Compiler with C++17 support is required` and the following
present in `CMakeError.log`:

```
Code Signing Warning: "cmTC_635cc" isn't code signed but requires entitlements. It is not possible to add entitlements to a binary without signing it.
Code Signing Error: Code signing is required for product type 'Command-line Tool' in SDK 'iOS 12.1'
```

refs #1097