chore: Optimize TestRunner build
Only do `xcodebuild` without performing the costy `xcodebuild archive`
and `xcodebuild exportArchive`. Make a zip file with the `TestRunner.app`
instead and call it `TestRunner.ipa`. We are only using this `.ipa` package
to extract the `.app` and run it on a device with `ios-deploy` so it really
doesn't need to be a "real" `.ipa`.