fix(runtime): Do not hide property with more accessors
A property can be defined from both a protocol and a base class.
Currently if it is in a protocol it takes precedence but there are
cases which the other way around would be preferable.

* (ObjCPrototype) Check for property descriptor of base class and return that we
don't define it if the base one has more accessors.
* (ObjCClassBuilder) When defining a property check for metadata with
and without considering protocols. Use the one with more accessors.
* Add unit tests