chore(build): Retry WebKit build when on Catalyst
The new build system has some caching mechanism which checks
in advance whether files to be compiled exist or not. It sporadically
fails to find `DerivedSources/WTF/mach_excServer.c` when the
generation is a bit slower on Jenkins slaves. As a workaround we're
retrying the `xcodebuild` command. The second time around the file
will have been generated in advance and no such racing condition will
exist.