fix(webkit): Do not show error window on protocol errors
* A lot of commands are not supported by NativeScript. We'd better
keep the previous behaviour of simply reporting without stopping
on such errors.

* Revive the commented `activateDomain` calls. Since they are no
longer shown in the UI its better to keep them, so that the frontend
starts working if they become implemented in the future.

* Match {N} CLI's debugger proxy port. This way it will be possible
to use Safari for debugging the frontend without relying on the
proxy in `src/debugging` of iOS Runtime's repository.