feat(metadata-generator): Filter according to specified black/white lists
Each text file contains a list of patterns of the form `moduleGlob:symbolGlob`,
where `?` and `*` special symbols treated as `any symbol` and `zero or
more symbols`.

A symbol is whitelisted if it's module name and native name pair matches a
pattern in the whitelist, or if a whitelist file is not passed as an argument. All
symbols which are not whitelisted or match a pattern in the blacklist are
excluded from the global table of metadata symbols.

* Update metadata-generator submodule
* Add a `blacklist.mdg` file to TestRunner and symlink from the root (because
SRCROOT of TestRunner points to the directory of the main CMakeLists.txt)
* Add a unit test which checks for the correct filtering of metadata according to
the rules in `blacklist.mdg`