test: Add cases with a blacklisted entities
* Add test case with function returning a struct which is blacklisted
If a struct's metadata is missing it's not possible to silently ignore it and
perform the call, because we don't know its size and cannot reliably
allocate the stack frame for the cif. That's why we'll abort the call and
throw a JS exception
* Add test case with functions returning and taking as an argument
objects from a blacklisted class. In this scenario the behavior is the
same as with objects from private classes
* Use Objective-C exceptions to propagate errors when filtered
metadata is used and handle them in JSC callbacks
* Update metadata generator to use native names in type encodings
to allow correct discovery of interfaces and protocols when they
have been filtered