fix: Register Swift classes with both native names
The Objective-C runtime doesn't behave consistently with Swift class
names. Sometimes the mangled name can only be used, other times
the demangled one. To avoid hardcoding any logic which may get
easily broken in the future, we are adding both names in the global
tables and are using the mangled name in signatures.

Observation which lead to this change:

* `objc_getClass(const char *)` - accepts both mangled and demangled
names of Swift classes which are not nested in another class. Nested
classes however can be discovered only by their mangled name.

* `class_getName` - returns the demangled name for Swift classes
which aren't nested, and the mangled one for nested.

* On Mac Catalyst `class_getName` returns  the mangled names
no matter whether the class is nested or not.

* Update the `metadata-generator` submodule
* Add an optional demangled name for every metadata entity
* Add a test case with a nested Swift-like class